+++ src/components/Description.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Description = () => {
-    return (<div className="div-card main-content">
+    return (<div className="main-content div-card">
         <p>This is a work in progress! Check the <a href="https://github.com/BionicCat27/My-Simple-Priorities-Web/blob/master/README.md">GitHub README</a> for more details on current and future features.</p>
         <p>Tip: Click on the priority panels for more options.</p>
     </div>);
