+++ src/components/Sidebar.js
@@ -2,13 +2,27 @@ import React from 'react';
 
 import './Sidebar.css';
 
+import '../firebaseConfig';
+
+import { getAuth, signOut } from "firebase/auth";
+
+const auth = getAuth();
+
 const Sidebar = () => {
+
+    function userSignOut() {
+        signOut(auth).then(() => {
+            window.location = " /login";
+        }).catch((error) => {
+            console.log("An error occurred during signout: " + error);
+        });
+    }
+
     return (
         <div id="sidebar">
             <div id="sidebarLinksContainer">
                 <a href="/">Home</a>
-                <a href="/">About</a>
-                <a href="/">Logout</a>
+                <a onClick={userSignOut}>Logout</a>
             </div>
             <div id="sidebarCreditContainer">
                 <a href="https://aslanbb.vercel.app/">Aslan Bennington-Barr</a>
